64bit-build: replace leftover long, unsigned long with int32_t/uint32_t

+++ src/tcp.h
@@ -8,13 +8,13 @@
 #ifndef tcp_H_
 #define tcp_H_
 
-extern long resolv(char [], unsigned long *);
+extern long resolv(char [], uint32_t *);
 
 extern long connectto(char [], char []);
 
 extern long waitconnect(char [], unsigned long);
 
-extern long acceptconnect(long, char [], long *);
+extern long acceptconnect(long, char [], uint32_t *);
 
 extern void ipnum2str(char [], char [], unsigned long);
 
@@ -24,9 +24,9 @@ extern long sendmore(long, char [], long);
 
 extern long readsock(long, char [], long);
 
-extern long getsockipnum(long, char [], long *);
+extern long getsockipnum(long, char [], uint32_t *);
 
-extern long getpeeripnum(long, char [], long *);
+extern long getpeeripnum(long, char [], uint32_t *);
 
 extern void ipnumport2str(char [], unsigned long, char [], unsigned long,
                 char [], unsigned long);
